% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_class.R
\name{check_class}
\alias{check_class}
\title{Checks that two objects have the same classes}
\usage{
check_class(
  object = .result,
  expected = .solution,
  unit = "result",
  problem_prefix = ""
)
}
\arguments{
\item{object}{A vector to be compared to \code{expected}.}

\item{expected}{A vector containing the expected result.}

\item{unit}{\verb{[character(1)]}\cr The label used to describe the object in
feedback messages. Defaults to \code{"result"}.}

\item{problem_prefix}{\verb{[character(1)]}\cr The prefix appended to the
\code{problem} label in \code{\link[gradethis:graded]{gradethis::fail()}} objects.
Defaults to \code{""}, which appends no prefix.}
}
\value{
If there are any issues, generate a \code{\link[gradethis:graded]{gradethis::fail()}} message.
Otherwise, invisibly return \code{\link{NULL}}.
}
\description{
Checks if \code{object} and \code{expected} have the same \link[=class]{class}.
If the classes differ, returns a failure state and an informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}.
}
\section{Problems}{

\enumerate{
\item \code{class}: \code{object} doesn't have the same classes as \code{expected}
}
}

