% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_column.R
\name{check_column}
\alias{check_column}
\title{Checks that a column is identical across two tables}
\usage{
check_column(
  name,
  object = .result,
  expected = .solution,
  check_class = TRUE,
  check_length = TRUE,
  check_values = TRUE,
  max_diffs = 3
)
}
\arguments{
\item{name}{\verb{[character(1)]}\cr The name of the column to check.}

\item{object}{A data frame to be compared to \code{expected}.}

\item{expected}{A data frame containing the expected result.}

\item{check_class}{\verb{[logical(1)]}\cr Whether to check that \code{name} has the
same class in \code{object} and \code{expected}.}

\item{check_length}{\verb{[logical(1)]}\cr Whether to check that \code{name} has the
same length in \code{object} and \code{expected}.}

\item{check_values}{\verb{[logical(1)]}\cr Whether to check that \code{name} has the
same values in \code{object} and \code{expected}.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of mismatched values to
print. Defaults to 3.}
}
\value{
If there are any issues, generate a \code{\link[gradethis:graded]{gradethis::fail()}} message.
Otherwise, invisibly return \code{\link{NULL}}.
}
\description{
Checks if the \code{name} column has the same class and values in \code{object} and
\code{expected}. If the columns differ, returns a failure state and an informative
message with \code{\link[gradethis:graded]{gradethis::fail()}}.
}
\section{Problems}{

\enumerate{
\item \code{column_class}: Any mismatch in the classes of the \code{name} column
\item \code{column_length}: The \code{name} column doesn't have the expected length
\item \code{column_values}: The \code{name} column doesn't have the expected values
\item \code{column_name}: The \code{name} column doesn't appear in the \code{object}
}
}

