% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{check_names}
\alias{check_names}
\title{Check that the names of two object are the same}
\usage{
check_names(
  object = .result,
  expected = .solution,
  max_diffs = 3,
  unit = "result",
  problem_prefix = ""
)
}
\arguments{
\item{object}{An object to be compared to \code{expected}.}

\item{expected}{An object containing the expected result.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of missing and/or
unexpected names to include in an informative failure message.
Defaults to 3.}

\item{unit}{\verb{[character(1)]}\cr The label used to describe the object in
feedback messages. Defaults to \code{"result"}.}

\item{problem_prefix}{\verb{[character(1)]}\cr The prefix appended to the
\code{problem} label in \code{\link[gradethis:graded]{gradethis::fail()}} objects.
Defaults to \code{""}, which appends no prefix.}
}
\value{
If there are any issues, generate a \code{\link[gradethis:graded]{gradethis::fail()}} message.
Otherwise, invisibly return \code{\link{NULL}}.
}
\description{
Checks if \code{object} and \code{expected} have the same \link[=names]{names}.
If the names differ, returns a failure state and an informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}.
}
\section{Problems}{

\enumerate{
\item \code{names}: The table has names that are not expected,
or is missing names that are expected.
}
}

