% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_table.R
\name{check_table}
\alias{check_table}
\title{Check that the rows and columns of two tables are the same}
\usage{
check_table(
  object = get(".result", parent.frame()),
  expected = get(".solution", parent.frame())
)
}
\arguments{
\item{object}{A data frame to be compared to \code{expected}.}

\item{expected}{A data frame containing the expected result.}
}
\value{
Invisible \code{\link{NULL}}
}
\description{
Checks if \code{object} and \code{expected} have the same number of rows, the same
column names, and the same column contents.
If the data frames differ, returns a failure state and an informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}.
}
