% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated-methods}
\alias{deprecated-methods}
\alias{tblcheck_grade}
\alias{tblcheck_message}
\title{Deprecated Generics}
\usage{
tblcheck_grade(problem, ...)

tblcheck_message(problem, ...)
}
\description{
These generics are now deprecated.
}
\section{Functions}{
\itemize{
\item \code{tblcheck_grade()}: is now \code{\link[=problem_grade]{problem_grade()}}.

\item \code{tblcheck_message()}: is now \code{\link[=problem_message]{problem_message()}}.

}}
\keyword{internal}
