% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_class-friendly_class.R
\name{friendly_class}
\alias{friendly_class}
\title{Generate a human-readable description of an object's class}
\usage{
friendly_class(object)
}
\arguments{
\item{object}{An object whose class will be described}
}
\value{
A \link{character} string of length 1
}
\description{
Generate a human-readable description of an object's class
}
