% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_this.R
\name{grade_this_vector}
\alias{grade_this_vector}
\title{Grade this vector}
\usage{
grade_this_vector(
  correct = NULL,
  pre_check = NULL,
  post_check = NULL,
  pass_if_equal = TRUE,
  ...,
  max_diffs = 3,
  check_class = TRUE,
  ignore_class = NULL,
  check_length = TRUE,
  check_levels = TRUE,
  check_values = TRUE,
  tolerance = sqrt(.Machine$double.eps),
  check_names = TRUE,
  hint = getOption("gradethis.fail.hint", FALSE),
  encourage = getOption("gradethis.fail.encourage", FALSE),
  pass.praise = NULL
)
}
\arguments{
\item{correct}{\verb{[character(1)]}\cr The message shown to the student when
their \code{.result} matches the exercise \code{.solution}, if \code{pass_if_equal} is
\code{TRUE}.}

\item{pre_check, post_check}{\verb{[expression]}\cr Code to run before or after the
table or vector grading is performed. The pre check runs before calling
\code{\link[gradethis:pass_if_equal]{gradethis::pass_if_equal()}} so that you can modify or adjust the student's
\code{.result} or the \code{.solution} if there are parts of either that need to be
ignored. These arguments can also be used in conjunction with the
\code{pass_if_equal} option when the grading requirements are more involved.}

\item{pass_if_equal}{\verb{[logical(1)]}\cr When \code{TRUE} (default for
\code{grade_this_vector()} but not \code{grade_this_table()}), the \code{.result} is
compared to the \code{.solution} with \code{\link[gradethis:pass_if_equal]{gradethis::pass_if_equal()}} after the
\emph{pre check} and before calling the \pkg{tblcheck} grading function.}

\item{...}{Additional arguments passed to \code{graded()} or additional data to be
included in the feedback object.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of mismatched values to
print. Defaults to 3.}

\item{check_class}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same classes.}

\item{ignore_class}{\verb{[character()]}\cr A vector of classes to ignore when
finding differences between \code{object} and \code{expected}.

If an element is named, differences will only be ignored between the pair
of the element and its name.
For example, \code{ignore_class = c("integer" = "numeric")} will ignore class
differences only if \code{object} has class \link{integer} and \code{expected} has class
\link{numeric}, or vice versa.

If all the classes of \code{expected} are included in \code{ignore_class},
a \code{class} problem will never be returned.}

\item{check_length}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same length.}

\item{check_levels}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same \link[=levels]{factor levels}.}

\item{check_values}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} contain the same values.}

\item{tolerance}{\verb{[numeric(1) ≥ 0]}\cr \code{values} differences smaller than
\code{tolerance} are ignored. The default value is close to \code{1.5e-8}.}

\item{check_names}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same names.}

\item{hint}{Include a code feedback hint with the failing message? This
argument only applies to \code{fail()} and \code{fail_if_equal()} and the message is
added using the default options of \code{\link[gradethis:give_code_feedback]{give_code_feedback()}} and
\code{\link[gradethis:maybe_code_feedback]{maybe_code_feedback()}}. The default value of \code{hint} can be set using
\code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.hint} option.}

\item{encourage}{Include a random encouraging phrase with
\code{\link[gradethis:random_encouragement]{random_encouragement()}}? The default value of \code{encourage} can be set
using \code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.encourage} option.}

\item{pass.praise}{Logical \code{TRUE} or \code{FALSE} to determine whether a praising
phrase should be automatically prepended to any \code{\link[gradethis:pass]{pass()}} or
\code{\link[gradethis:pass_if_equal]{pass_if_equal()}} messages. Sets the \code{gradethis.pass.praise} option.}
}
\value{
The returned feedback is equivalent to \pkg{gradethis} grading code
using \code{\link[gradethis:grade_this]{grade_this()}} with the following
components:
\enumerate{
\item First the \code{pre_check} code, if any, is evaluated. If this code calls
\code{\link[gradethis:graded]{pass()}}, \code{\link[gradethis:graded]{fail()}}, or their
equivalents, that feedback is provided immediately.
\item If \code{pass_if_equal} is \code{TRUE}, then
\code{\link[gradethis:pass_if_equal]{pass_if_equal()}} is called to compare the
\code{\link[gradethis:grade_this-objects]{.result}} to the
\code{\link[gradethis:grade_this-objects]{.solution}}. The message in \code{correct}
is used for the feedback.
\item The appropriate \pkg{tblcheck} grading function is called, returning
any feedback:
\enumerate{
\item \code{grade_this_table()} returns the results from \code{\link[=tbl_grade]{tbl_grade()}}
\item \code{grade_this_vector()} returns the results from \code{\link[=vec_grade]{vec_grade()}}
}
\item The \code{post_check} code, if any, is evaluated and any feedback from a call
to \code{\link[gradethis:graded]{pass()}}, \code{\link[gradethis:graded]{fail()}}, or
their equivalents is returned.
\item Finally, if no other feedback is returned, the feedback from
\code{\link[gradethis:fail]{gradethis::fail()}} is provided to the student, using the options
\code{fail.message}, \code{fail.hint} and \code{fail.encourage}.
}
}
\description{
Automatically grade a vector resulting from student code using
\code{\link[gradethis:grade_this]{gradethis::grade_this()}} and \code{\link[=vec_grade]{vec_grade()}} to compare the
student's result with the author's solution.
}
\examples{
ex <- gradethis::mock_this_exercise(
	.solution_code = tibble::tibble(x = 1:3, y = letters[x]),
	.user_code = tibble::tibble(x = 1:3, y = c("A", "b", "c"))
)

#' ## Grading Vectors ----
# Here we use `pre_check` to modify `.result` and
grade_this_vector(
	pre_check = {
		.result <- .result$y
		.solution <- .solution$y
	}
)(ex)

# Roughly equivalent to...
gradethis::grade_this({
	.result <- .result$y
	.solution <- .solution$y
	gradethis::pass_if_equal()
	vec_grade()
	gradethis::fail()
})(ex)
}
\seealso{
\code{\link[=vec_grade]{vec_grade()}}

Other graders:
\code{\link[=grade_this_table]{grade_this_table()}}
}
\concept{graders}
