% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblcheck.R
\name{has_classes}
\alias{has_classes}
\title{Checks if a table has expected classes.}
\usage{
has_classes(x, reference = NULL, classes = NULL, exact = TRUE)
}
\arguments{
\item{x}{tibble or data.frame}

\item{reference}{tibble or data.frame}

\item{classes}{character}

\item{exact}{A boolean to indicate whether to use exact matching}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Checks if a table has expected classes.
}
\examples{
has_classes(mtcars, mtcars)
has_classes(mtcars, classes = "data.frame")
}
