% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{has_column_types}
\alias{has_column_types}
\title{Checks if a table has the right column types.}
\usage{
has_column_types(x, reference = NULL, col_types = NULL)
}
\arguments{
\item{x}{data.frame or tibble}

\item{reference}{data.frame or tibble}

\item{col_types}{character}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Checks if a table has the right column types.
}
\examples{
has_column_types(mtcars, reference = mtcars)
}
