% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{has_extra_columns}
\alias{has_extra_columns}
\title{Checks if a table has extra columns.}
\usage{
has_extra_columns(x, reference = NULL, col_names = NULL)
}
\arguments{
\item{x}{data.frame or tibble}

\item{reference}{data.frame or tibble}

\item{col_names}{character}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Checks if a table has extra columns.
}
\examples{
has_extra_columns(mtcars, mtcars)
has_extra_columns(mtcars, col_names = "disp")
}
