% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{has_missing_columns}
\alias{has_missing_columns}
\title{Checks if a table has missing columns.}
\usage{
has_missing_columns(x, reference = NULL, col_names = NULL)
}
\arguments{
\item{x}{data.frame or tibble}

\item{reference}{data.frame or tibble}

\item{col_names}{character}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Checks if a table has missing columns.
}
\examples{
has_missing_columns(mtcars[, 1:2], mtcars)
has_missing_columns(mtcars[, 1:2], col_names = names(mtcars))
}
