% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{has_n_cols}
\alias{has_n_cols}
\title{Checks if a table has the expected number of columns.}
\usage{
has_n_cols(x, reference = NULL, n_cols = NULL)
}
\arguments{
\item{x}{tibble or data.frame}

\item{reference}{tibble or data.frame}

\item{n_cols}{numeric}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Checks if a table has the expected number of columns.
}
\examples{
has_n_cols(mtcars, mtcars)
has_n_cols(mtcars, n_cols = 11)
}
