% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{has_n_rows}
\alias{has_n_rows}
\title{Checks if a table has the expected number of rows.}
\usage{
has_n_rows(x, reference = NULL, n_rows = NULL)
}
\arguments{
\item{x}{tibble or data.frame}

\item{reference}{tibble or data.frame}

\item{n_rows}{numeric}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Checks if a table has the expected number of rows.
}
\examples{
has_n_rows(mtcars, mtcars)
has_n_rows(mtcars, n_rows = 32)
}
