% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_class-hinted_class_message.R
\name{hinted_class_message}
\alias{hinted_class_message}
\title{Generate a hint for how to convert one object type to another}
\usage{
hinted_class_message(object, expected)
}
\arguments{
\item{object}{An object to be compared to \code{expected}}

\item{expected}{An object of the expected class}
}
\value{
A \link{character} string of length 1
}
\description{
Generate a hint for how to convert one object type to another
}
