% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{list_setdiff}
\alias{list_setdiff}
\title{List the difference between two vectors in a user-friendly string}
\usage{
list_setdiff(x, y, max = Inf, ...)
}
\arguments{
\item{x}{Vectors. Elements that appear in \code{x} but not \code{y} will be listed.}

\item{y}{Vectors. Elements that appear in \code{x} but not \code{y} will be listed.}

\item{max}{The maximum number of elements to list.
Defaults to \link{Inf}, which includes all elements.}

\item{...}{Additional arguments passed to \code{\link[knitr:combine_words]{knitr::combine_words()}}}
}
\value{
A single \link{character} string.
}
\description{
Calls \code{\link[=setdiff]{setdiff()}}, optionally limits the result to \code{max} items, and then
combines the result with \code{\link[knitr:combine_words]{knitr::combine_words()}}.
}
\keyword{internal}
