% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{max_setdiff}
\alias{max_setdiff}
\title{Find up to a certain number of differences between two vectors}
\usage{
max_setdiff(
  x,
  y,
  max_diffs = Inf,
  and = " and ",
  column = FALSE,
  transform = md_code
)
}
\arguments{
\item{x, y}{Vectors. Elements that appear in \code{x} but not \code{y} will be listed.}

\item{max_diffs}{The maximum number of elements to list.
Defaults to \link{Inf}, which includes all elements.}

\item{and}{Character string to be prepended to the last word.}

\item{transform}{A function applied to each element of the items appearing in
the string, prior to combining into a string. Use \code{identity} to keep the
items without modification.}

\item{unit}{An optional string describing the elements in \code{x} and \code{y}. Passed
to \code{\link[plu:plu_ral]{plu::ral()}}. Should be written in the singular form, e.g. \code{"n column"}.}
}
\value{
A list with three items:
\itemize{
\item \code{diffs} contains the full \code{setdiff()} between \code{x} and \code{y}.
\item \code{str} contains a string describing the differences up to \code{max_diffs}.
\item \code{unit} contains the pluralized \code{unit}, relative to the number of \code{diffs}.
}
}
\description{
Find up to a certain number of differences between two vectors
}
\keyword{internal}
