% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_grade.R
\name{problem_grade}
\alias{problem_grade}
\alias{problem_grade.default}
\alias{problem_grade.list}
\alias{problem_grade.gradethis_problem}
\alias{problem_grade.tblcheck_problem}
\title{Apply automatic grading to a problem object}
\usage{
problem_grade(problem, max_diffs = 3, env = parent.frame(), ...)

\method{problem_grade}{default}(problem, max_diffs = 3, env = parent.frame(), ...)

\method{problem_grade}{list}(problem, max_diffs = 3, env = parent.frame(), ...)

\method{problem_grade}{gradethis_problem}(problem, max_diffs = 3, env = parent.frame(), ...)

\method{problem_grade}{tblcheck_problem}(problem, max_diffs = 3, env = parent.frame(), ...)
}
\arguments{
\item{problem}{A problem generated by \code{\link[=tbl_check]{tbl_check()}}, \code{\link[=vec_check]{vec_check()}} or their
related helper functions.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of mismatched values to
display in an informative failure message.
Passed to \code{\link[=tbl_check_names]{tbl_check_names()}} to determine the number of mismatched column
names to display and the \code{n_values} argument of \code{\link[=tbl_check_column]{tbl_check_column()}} to
determine the number of mismatched column values to display.
Defaults to 3.}

\item{env}{The environment used for grading.}

\item{...}{
  Arguments passed on to \code{\link[gradethis:fail]{gradethis::fail}}
  \describe{
    \item{\code{correct}}{A logical value of whether or not the checked code is correct.}
    \item{\code{type,location}}{The \code{type} and \code{location} of the feedback object
provided to \pkg{learnr}. See
\url{https://rstudio.github.io/learnr/exercises.html#Custom_checking} for more
details.

\code{type} may be one of "auto", "success", "info", "warning", "error", or
"custom".

\code{location} may be one of "append", "prepend", or "replace".}
    \item{\code{praise}}{Include a random praising phrase with \code{\link[gradethis:random_praise]{random_praise()}}? The
default value of \code{praise} can be set using \code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the
\code{gradethis.pass.praise} option.}
    \item{\code{hint}}{Include a code feedback hint with the failing message? This
argument only applies to \code{fail()} and \code{fail_if_equal()} and the message is
added using the default options of \code{\link[gradethis:give_code_feedback]{give_code_feedback()}} and
\code{\link[gradethis:maybe_code_feedback]{maybe_code_feedback()}}. The default value of \code{hint} can be set using
\code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.hint} option.}
    \item{\code{encourage}}{Include a random encouraging phrase with
\code{\link[gradethis:random_encouragement]{random_encouragement()}}? The default value of \code{encourage} can be set
using \code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.encourage} option.}
  }}
}
\value{
A \code{\link[gradethis:fail]{gradethis::fail()}} message or \code{NULL} invisibly.
}
\description{
Automatically converts a \code{\link[=problem]{problem()}} object into a \pkg{gradethis} grade.
\code{problem_grade()} is an S4 generic and \pkg{tblcheck} provides an internal
method for problems with class \code{"tblcheck_problem"}. In \pkg{tblcheck}, or
for problems with this class, any problems are automatically turned into
failing grades with \code{\link[gradethis:fail]{gradethis::fail()}} and using the message provided by
\code{\link[=problem_message]{problem_message()}}.
}
\examples{
.result <- 1:10
.solution <- letters[1:10]
problem <- vec_check()
problem_grade(problem)

}
\seealso{
Other Problem functions:
\code{\link[=problem]{problem()}},
\code{\link[=problem_message]{problem_message()}},
\code{\link[=problem_type]{problem_type()}}
}
\concept{Problem functions}
