% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{.result}
\alias{.solution}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gradethis}{\code{\link[gradethis:.result]{.result}}, \code{\link[gradethis:.solution]{.solution}}}
}}

