% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_check.R
\name{tbl_check}
\alias{tbl_check}
\alias{tbl_grade}
\title{Check that the rows and columns of two tables are the same}
\usage{
tbl_check(
  object = .result,
  expected = .solution,
  cols = NULL,
  check_class = TRUE,
  ignore_class = NULL,
  check_names = TRUE,
  check_column_order = FALSE,
  check_dimensions = TRUE,
  check_groups = TRUE,
  check_columns = TRUE,
  check_column_class = check_columns,
  check_column_levels = check_columns,
  check_column_values = check_columns,
  tolerance = sqrt(.Machine$double.eps),
  check_row_order = check_columns,
  env = parent.frame()
)

tbl_grade(
  object = .result,
  expected = .solution,
  cols = NULL,
  max_diffs = 3,
  check_class = TRUE,
  ignore_class = NULL,
  check_names = TRUE,
  check_column_order = FALSE,
  check_dimensions = TRUE,
  check_groups = TRUE,
  check_columns = TRUE,
  check_column_class = check_columns,
  check_column_levels = check_columns,
  check_column_values = check_columns,
  tolerance = sqrt(.Machine$double.eps),
  check_row_order = check_columns,
  env = parent.frame(),
  ...
)
}
\arguments{
\item{object}{A data frame to be compared to \code{expected}.}

\item{expected}{A data frame containing the expected result.}

\item{cols}{[\code{\link[tidyselect:language]{tidy-select}}]\cr A selection of columns
to compare between \code{object} and \code{expected}. Differences in other columns
will be ignored. If \code{\link{NULL}}, the default, all columns will be checked.}

\item{check_class}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same classes with \code{\link[=tbl_check_class]{tbl_check_class()}}.}

\item{ignore_class}{\verb{[character()]}\cr A vector of classes to ignore when
finding differences between \code{object} and \code{expected}.

If an element is named, differences will only be ignored between the pair
of the element and its name.
For example, \code{ignore_class = c("integer" = "numeric")} will ignore class
differences only if \code{object} has class \link{integer} and \code{expected} has class
\link{numeric}, or vice versa.

If all the classes of \code{expected} are included in \code{ignore_class},
a \code{class} problem will never be returned.}

\item{check_names}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same column names with \code{\link[=tbl_check_names]{tbl_check_names()}}.}

\item{check_column_order}{\verb{[logical(1)]}\cr Whether to check that the columns
of  \code{object} are in the same order as \code{expected} with \code{\link[=tbl_check_names]{tbl_check_names()}}.
Defaults to \code{FALSE}.}

\item{check_dimensions}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same number of rows and columns
with \code{\link[=tbl_check_dimensions]{tbl_check_dimensions()}}.}

\item{check_groups}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same \link[dplyr:group_by]{groups}
with \code{\link[dplyr:group_vars]{dplyr::group_vars()}}.}

\item{check_columns}{\verb{[logical(1)]}\cr Whether to check that all columns
have the same contents with \code{\link[=tbl_check_column]{tbl_check_column()}}.}

\item{check_column_class}{\verb{[logical(1)]}\cr Whether to check that each
column has the same class in \code{object} and \code{expected}.}

\item{check_column_levels}{\verb{[logical(1)]}\cr Whether to check that each
column has the same \link[=levels]{factor levels} in \code{object} and \code{expected}.}

\item{check_column_values}{\verb{[logical(1)]}\cr Whether to check that each
column has the same values in \code{object} and \code{expected}.}

\item{tolerance}{\verb{[numeric(1) ≥ 0]}\cr \code{values} differences smaller than
\code{tolerance} are ignored. The default value is close to \code{1.5e-8}.}

\item{check_row_order}{\verb{[logical(1)]}\cr Whether to check that the values in
each column are in the same order in \code{object} and \code{expected}.}

\item{env}{The environment in which to find \code{.result} and \code{.solution}.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of mismatched values to
display in an informative failure message.
Passed to \code{\link[=tbl_check_names]{tbl_check_names()}} to determine the number of mismatched column
names to display and the \code{n_values} argument of \code{\link[=tbl_check_column]{tbl_check_column()}} to
determine the number of mismatched column values to display.
Defaults to 3.}

\item{...}{
  Arguments passed on to \code{\link[gradethis:fail]{gradethis::fail}}
  \describe{
    \item{\code{correct}}{A logical value of whether or not the checked code is correct.}
    \item{\code{type,location}}{The \code{type} and \code{location} of the feedback object
provided to \pkg{learnr}. See
\url{https://rstudio.github.io/learnr/exercises.html#Custom_checking} for more
details.

\code{type} may be one of "auto", "success", "info", "warning", "error", or
"custom".

\code{location} may be one of "append", "prepend", or "replace".}
    \item{\code{praise}}{Include a random praising phrase with \code{\link[gradethis:random_praise]{random_praise()}}? The
default value of \code{praise} can be set using \code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the
\code{gradethis.pass.praise} option.}
    \item{\code{hint}}{Include a code feedback hint with the failing message? This
argument only applies to \code{fail()} and \code{fail_if_equal()} and the message is
added using the default options of \code{\link[gradethis:give_code_feedback]{give_code_feedback()}} and
\code{\link[gradethis:maybe_code_feedback]{maybe_code_feedback()}}. The default value of \code{hint} can be set using
\code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.hint} option.}
    \item{\code{encourage}}{Include a random encouraging phrase with
\code{\link[gradethis:random_encouragement]{random_encouragement()}}? The default value of \code{encourage} can be set
using \code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.encourage} option.}
  }}
}
\value{
If there are any issues, a \link{list} from \code{tbl_check()} or a
\code{\link[gradethis:fail]{gradethis::fail()}} message from \code{tbl_grade()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks for differences between \code{object} and \code{expected} in the following order:
\enumerate{
\item Check table class with \code{\link[=tbl_check_class]{tbl_check_class()}}
\item Check column names with \code{\link[=tbl_check_names]{tbl_check_names()}}
\item Check number of rows and columns with \code{\link[=tbl_check_dimensions]{tbl_check_dimensions()}}
\item Check \link[dplyr:group_by]{groups} with \code{\link[=tbl_check_groups]{tbl_check_groups()}}
\item Check that each column is the same with \code{\link[=tbl_check_column]{tbl_check_column()}}
}

If the tables differ
\itemize{
\item \code{tbl_check()} returns a list describing the problem
\item \code{tbl_grade()} returns a failing grade and informative message
with \code{\link[gradethis:fail]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{class}: The table does not have the expected classes.
\item \code{not_table}: \code{object} does not inherit the \link{data.frame} class.
\item \code{names}: The table has column names that are not expected,
or is missing names that are expected.
\item \code{names_order}: The table has the same column names as expected,
but in a different order.
\item \code{ncol}: The table doesn't have the expected number of columns.
\item \code{nrow}: The table doesn't have the expected number of rows.
\item \code{groups}: The table has \link[dplyr:group_by]{groups} that are
not expected, or is missing groups that are expected.
}

Additional problems may be produced by \code{\link[=tbl_check_column]{tbl_check_column()}}.
}

\examples{
.result <- data.frame(a = 1:10, b = 11:20)
.solution <- tibble::tibble(a = 1:10, b = 11:20)
tbl_check()
tbl_grade()

.result <- tibble::tibble(a = 1:10, b = a, c = a, d = a, e = a, f = a)
.solution <- tibble::tibble(z = 1:10, y = z, x = z, w = z, v = z, u = z)
tbl_check()
tbl_grade()
tbl_grade(max_diffs = 5)
tbl_grade(max_diffs = Inf)

.result <- tibble::tibble(a = 1:10, b = 11:20)
.solution <- tibble::tibble(a = 1:11, b = 12:22)
tbl_check()
tbl_grade()

.result <- tibble::tibble(a = 1:10, b = 11:20)
.solution <- tibble::tibble(a = letters[1:10], b = letters[11:20])
tbl_check()
tbl_grade()

.result <- tibble::tibble(a = 1:10, intermediate = 6:15, b = 11:20)
.solution <- tibble::tibble(a = 1:10, b = 11:20)
tbl_check(cols = any_of(names(.solution)))
tbl_grade(cols = any_of(names(.solution)))

.result <- tibble::tibble(a = 1:10, b = 11:20)
.solution <- tibble::tibble(a = 11:20, b = 1:10)
tbl_check()
tbl_grade()
tbl_grade(max_diffs = 5)
tbl_grade(max_diffs = Inf)

.result <- tibble::tibble(a = 1:10, b = rep(1:2, 5))
.solution <- dplyr::group_by(tibble::tibble(a = 1:10, b = rep(1:2, 5)), b)
tbl_check()
tbl_grade()
tbl_grade(check_groups = FALSE)
}
