% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_check_is_table.R
\name{tbl_check_is_table}
\alias{tbl_check_is_table}
\alias{tbl_grade_is_table}
\title{Checks that an object is a table}
\usage{
tbl_check_is_table(object = .result, env = parent.frame())

tbl_grade_is_table(object = .result, env = parent.frame(), ...)
}
\arguments{
\item{object}{An object to be compared to \code{expected}.}

\item{env}{The environment in which to find \code{.result} and \code{.solution}.}

\item{...}{
  Arguments passed on to \code{\link[gradethis:fail]{gradethis::fail}}
  \describe{
    \item{\code{correct}}{A logical value of whether or not the checked code is correct.}
    \item{\code{type,location}}{The \code{type} and \code{location} of the feedback object
provided to \pkg{learnr}. See
\url{https://rstudio.github.io/learnr/exercises.html#Custom_checking} for more
details.

\code{type} may be one of "auto", "success", "info", "warning", "error", or
"custom".

\code{location} may be one of "append", "prepend", or "replace".}
    \item{\code{praise}}{Include a random praising phrase with \code{\link[gradethis:random_praise]{random_praise()}}? The
default value of \code{praise} can be set using \code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the
\code{gradethis.pass.praise} option.}
    \item{\code{hint}}{Include a code feedback hint with the failing message? This
argument only applies to \code{fail()} and \code{fail_if_equal()} and the message is
added using the default options of \code{\link[gradethis:give_code_feedback]{give_code_feedback()}} and
\code{\link[gradethis:maybe_code_feedback]{maybe_code_feedback()}}. The default value of \code{hint} can be set using
\code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.hint} option.}
    \item{\code{encourage}}{Include a random encouraging phrase with
\code{\link[gradethis:random_encouragement]{random_encouragement()}}? The default value of \code{encourage} can be set
using \code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.encourage} option.}
  }}
}
\value{
If there are any issues, a \link{list} from \code{tbl_check_is_table()} or a
\code{\link[gradethis:fail]{gradethis::fail()}} message from \code{tbl_grade_is_table()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks if \code{object} inherits the \link{data.frame} class.
If the not
\itemize{
\item \code{tbl_check_is_table()} returns a list describing the problem
\item \code{tbl_grade_is_table()} returns a failing grade and
informative message with \code{\link[gradethis:fail]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{not_table}: The object is not a table
}
}

\examples{
.result <- data.frame(a = 1:10)
tbl_check_is_table()
tbl_grade_is_table()

.result <- tibble::tibble(a = 1:10)
tbl_check_is_table()
tbl_grade_is_table()

.result <- list(a = 1:10)
tbl_check_is_table()
tbl_grade_is_table()
}
