% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_length.R
\name{tbl_check_length}
\alias{tbl_check_length}
\alias{tbl_grade_length}
\title{Check that the length of two object are the same}
\usage{
tbl_check_length(
  object = .result,
  expected = .solution,
  dimension = c("length", "ncol", "nrow"),
  envir = parent.frame()
)

tbl_grade_length(
  object = .result,
  expected = .solution,
  dimension = c("length", "ncol", "nrow"),
  envir = parent.frame()
)
}
\arguments{
\item{object}{An object to be compared to \code{expected}.}

\item{expected}{An object containing the expected result.}

\item{dimension}{\verb{[character(1)]}\cr The dimension on which to compare the
lengths of \code{object} and \code{expected}.
One of \code{"length"}, \code{"ncol"}, or "\code{nrow}".
Defaults to \code{"length"}.}

\item{envir}{The environment in which to find \code{.result} and \code{.solution}.}
}
\value{
If there are any issues, a \link{list} from \code{tbl_check_length()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{tbl_grade_length()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks if \code{object} and \code{expected} have the same \link[=length]{length}.
If the names differ
\itemize{
\item \code{tbl_check_names()} returns a list describing the problem
\item \code{tbl_grade_names()} returns a failing grade and informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{length}: \code{object} and \code{expected} have different lengths.
\item \code{ncol}: \code{object} and \code{expected} have a different number of columns.
\item \code{nrow}: \code{object} and \code{expected} have a different number of rows
}
}

