% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{tbl_check_missing}
\alias{tbl_check_missing}
\alias{tbl_grade_missing}
\title{Checks that a column is identical across two tables}
\usage{
tbl_check_missing(column, object = .result, env = parent.frame())

tbl_grade_missing(column, object = .result, env = parent.frame())
}
\arguments{
\item{column}{\verb{[character(1)]}\cr The name of the column to check.}

\item{object}{A data frame to be compared to \code{expected}.}

\item{env}{The environment in which to find \code{.result} and \code{.solution}.}
}
\value{
If there are any issues, a \link{list} from \code{tbl_check_missing()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{tbl_grade_missing()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks if \code{column} exists in \code{object}.
If the column does not exist
\itemize{
\item \code{tbl_check_missing()} returns a list describing the problem
\item \code{tbl_grade_missing()} returns a failing grade and informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{missing}: The \code{column} doesn't appear in the \code{object}
}
}

\examples{
.result <- tibble::tibble(a = 1:10, b = 11:20)
.solution <- tibble::tibble(a = 1:10, b = 11:20)
tbl_check_missing("b")
tbl_grade_missing("b")
}
