% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_table.R
\name{tbl_check_table}
\alias{tbl_check_table}
\alias{tbl_grade_table}
\title{Check that the rows and columns of two tables are the same}
\usage{
tbl_check_table(
  object = .result,
  expected = .solution,
  max_diffs = 3,
  check_class = TRUE,
  check_nrow = TRUE,
  check_names = TRUE,
  check_ncol = !check_names,
  check_columns = TRUE,
  check_column_class = check_columns,
  check_column_values = check_columns,
  envir = parent.frame()
)

tbl_grade_table(
  object = .result,
  expected = .solution,
  max_diffs = 3,
  check_class = TRUE,
  check_nrow = TRUE,
  check_names = TRUE,
  check_ncol = !check_names,
  check_columns = TRUE,
  check_column_class = check_columns,
  check_column_values = check_columns,
  envir = parent.frame()
)
}
\arguments{
\item{object}{A data frame to be compared to \code{expected}.}

\item{expected}{A data frame containing the expected result.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of mismatched values to
display in an informative failure message.
Passed to \code{\link[=tbl_check_names]{tbl_check_names()}} to determine the number of mismatched column
names to display and the \code{n_values} argument of \code{\link[=tbl_check_column]{tbl_check_column()}} to
determine the number of mismatched column values to display.
Defaults to 3.}

\item{check_class}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same classes with \code{\link[=class]{class()}}.}

\item{check_nrow}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same number of rows with \code{\link[=nrow]{nrow()}}.}

\item{check_names}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same column names with \code{\link[=tbl_check_names]{tbl_check_names()}}.}

\item{check_ncol}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same number of columns with \code{\link[=ncol]{ncol()}}.
By default, \code{check_ncol} is \code{\link{FALSE}} if \code{check_names} is \code{\link{TRUE}}, as
running both is redundant.}

\item{check_columns}{\verb{[logical(1)]}\cr Whether to check that all columns
have the same contents with \code{\link[=tbl_check_column]{tbl_check_column()}}.}

\item{check_column_class}{\verb{[logical(1)]}\cr Whether to check that each
columns has the same class in \code{object} and \code{expected}.}

\item{check_column_values}{\verb{[logical(1)]}\cr Whether to check that each
column has the same values in \code{object} and \code{expected}.}

\item{envir}{The environment in which to find \code{.result} and \code{.solution}.}
}
\value{
If there are any issues, a \link{list} from \code{tbl_check_table()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{tbl_grade_table()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks if \code{object} and \code{expected} have the same number of rows, the same
column names, and the same column contents.
If the tables differ
\itemize{
\item \code{tbl_check_names()} returns a list describing the problem
\item \code{tbl_grade_names()} returns a failing grade and informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{table_class}: The table does not have the expected classes
\item \code{table_nrow}: The table doesn't have the expected number of rows
\item \code{table_ncol}: The table doesn't have the expected number of columns
\item \code{table_names}: The table has names that are not expected,
or is missing names that are expected.
}

Additional problems may be produced by \code{\link[=tbl_check_column]{tbl_check_column()}}
}

