% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{tbl_grade}
\alias{tbl_grade}
\title{Apply automatic grading to a problem object}
\usage{
tbl_grade(problem, max_diffs = 3)
}
\arguments{
\item{problem}{A problem generated by a \verb{tbl_check_*()} function.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of mismatched values to
display in an informative failure message.
Passed to \code{\link[=tbl_check_names]{tbl_check_names()}} to determine the number of mismatched column
names to display and the \code{n_values} argument of \code{\link[=tbl_check_column]{tbl_check_column()}} to
determine the number of mismatched column values to display.
Defaults to 3.}
}
\value{
A \code{\link[gradethis:graded]{gradethis::fail()}} message.
}
\description{
Apply automatic grading to a problem object
}
\examples{
.result <- 1:10
.solution <- letters[1:10]
problem <- tbl_check_vector()
tbl_grade(problem)
}
