% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblcheck-package.R
\docType{package}
\name{tblcheck-package}
\alias{tblcheck}
\alias{tblcheck-package}
\title{tblcheck: Grade Tables in Learning Exercises}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

'tblcheck' extends 'gradethis' with functions that inspect data frame or tibble objects to make it easier for teachers to check that student tables meet expectations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkgs.rstudio.com/tblcheck}
  \item \url{https://github.com/rstudio/tblcheck}
  \item Report bugs at \url{https://github.com/rstudio/tblcheck/issues}
}

}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alex.rossellhayes@rstudio.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID})

Authors:
\itemize{
  \item Alexander Rossell Hayes \email{alex.rossellhayes@rstudio.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID})
  \item Garrick Aden-Buie \email{garrick@rstudio.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})
}

Other contributors:
\itemize{
  \item Sara Altman \email{sara.altman@rstudio.com} (\href{https://orcid.org/0000-0002-2529-5680}{ORCID}) [contributor]
  \item  RStudio, PBC [copyright holder, funder]
}

}
\keyword{internal}
