% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_check.R
\name{vec_check}
\alias{vec_check}
\alias{vec_grade}
\title{Checks that two vectors are the same}
\usage{
vec_check(
  object = .result,
  expected = .solution,
  check_class = TRUE,
  ignore_class = NULL,
  check_length = TRUE,
  check_levels = TRUE,
  check_values = TRUE,
  tolerance = sqrt(.Machine$double.eps),
  check_names = TRUE,
  env = parent.frame()
)

vec_grade(
  object = .result,
  expected = .solution,
  max_diffs = 3,
  check_class = TRUE,
  ignore_class = NULL,
  check_length = TRUE,
  check_levels = TRUE,
  check_values = TRUE,
  tolerance = sqrt(.Machine$double.eps),
  check_names = TRUE,
  env = parent.frame(),
  ...
)
}
\arguments{
\item{object}{A vector to be compared to \code{expected}.}

\item{expected}{A vector containing the expected result.}

\item{check_class}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same classes.}

\item{ignore_class}{\verb{[character()]}\cr A vector of classes to ignore when
finding differences between \code{object} and \code{expected}.

If an element is named, differences will only be ignored between the pair
of the element and its name.
For example, \code{ignore_class = c("integer" = "numeric")} will ignore class
differences only if \code{object} has class \link{integer} and \code{expected} has class
\link{numeric}, or vice versa.

If all the classes of \code{expected} are included in \code{ignore_class},
a \code{class} problem will never be returned.}

\item{check_length}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same length.}

\item{check_levels}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same \link[=levels]{factor levels}.}

\item{check_values}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} contain the same values.}

\item{tolerance}{\verb{[numeric(1) ≥ 0]}\cr \code{values} differences smaller than
\code{tolerance} are ignored. The default value is close to \code{1.5e-8}.}

\item{check_names}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same names.}

\item{env}{The environment in which to find \code{.result} and \code{.solution}.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of mismatched values to
print. Defaults to 3.}

\item{...}{
  Arguments passed on to \code{\link[gradethis:fail]{gradethis::fail}}
  \describe{
    \item{\code{correct}}{A logical value of whether or not the checked code is correct.}
    \item{\code{type,location}}{The \code{type} and \code{location} of the feedback object
provided to \pkg{learnr}. See
\url{https://rstudio.github.io/learnr/exercises.html#Custom_checking} for more
details.

\code{type} may be one of "auto", "success", "info", "warning", "error", or
"custom".

\code{location} may be one of "append", "prepend", or "replace".}
    \item{\code{praise}}{Include a random praising phrase with \code{\link[gradethis:random_praise]{random_praise()}}? The
default value of \code{praise} can be set using \code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the
\code{gradethis.pass.praise} option.}
    \item{\code{hint}}{Include a code feedback hint with the failing message? This
argument only applies to \code{fail()} and \code{fail_if_equal()} and the message is
added using the default options of \code{\link[gradethis:give_code_feedback]{give_code_feedback()}} and
\code{\link[gradethis:maybe_code_feedback]{maybe_code_feedback()}}. The default value of \code{hint} can be set using
\code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.hint} option.}
    \item{\code{encourage}}{Include a random encouraging phrase with
\code{\link[gradethis:random_encouragement]{random_encouragement()}}? The default value of \code{encourage} can be set
using \code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.encourage} option.}
  }}
}
\value{
If there are any issues, a \link{list} from \code{vec_check()} or a
\code{\link[gradethis:fail]{gradethis::fail()}} message from \code{vec_grade()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks for differences between \code{object} and \code{expected} in the following order:
\enumerate{
\item Check class with \code{\link[=vec_check_class]{vec_check_class()}}
\item Check length with \code{\link[=vec_check_dimensions]{vec_check_dimensions()}}
\item If the vector is a factor, check factor levels are the same with \code{\link[=vec_check_levels]{vec_check_levels()}}
\item Check vector values are the same with \code{\link[=vec_check_values]{vec_check_values()}}
\item Check names with \code{\link[=vec_check_names]{vec_check_names()}}
}

If the vectors differ
\itemize{
\item \code{vec_check()} returns a list describing the problem
\item \code{vec_grade()} returns a failing grade and informative message
with \code{\link[gradethis:fail]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{class}: \code{object} doesn't have the same classes as \code{expected}.
\item \code{length}: \code{object} doesn't have the same length as \code{expected}.
\item \code{levels_n}, \code{levels}, \code{levels_reversed}, \code{levels_order}:
See \code{\link[=vec_check_levels]{vec_check_levels()}}.
\item \code{values}: \code{object} doesn't contain the same values as \code{expected}.
\item \code{names}: \code{object} has different \link{names} than \code{expected}.
\item \code{names_order}: \code{object} has the same \link{names} as \code{expected},
but in a different order.
}
}

\examples{
.result <- 1:10
.solution <- letters[1:10]
vec_check()
vec_grade()

.result <- 1:10
.solution <- 1:11
vec_check()
vec_grade()

.result <- 1:10
.solution <- rlang::set_names(1:10, letters[1:10])
vec_check()
vec_grade()
vec_grade(max_diffs = 5)
vec_grade(max_diffs = Inf)

.result <- 1:10
.solution <- 11:20
vec_check()
vec_grade()
vec_grade(max_diffs = 5)
vec_grade(max_diffs = Inf)
}
