% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_length.R
\name{vec_check_length}
\alias{vec_check_length}
\alias{vec_grade_length}
\title{Check that the length of two object are the same}
\usage{
vec_check_length(object = .result, expected = .solution, env = parent.frame())

vec_grade_length(object = .result, expected = .solution, env = parent.frame())
}
\arguments{
\item{object}{An object to be compared to \code{expected}.}

\item{expected}{An object containing the expected result.}

\item{env}{The environment in which to find \code{.result} and \code{.solution}.}
}
\value{
If there are any issues, a \link{list} from \code{vec_check_length()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{vec_grade_length()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks if \code{object} and \code{expected} have the same \link[=length]{length}.
If the lengths differ
\itemize{
\item \code{vec_check_lengths()} returns a list describing the problem
\item \code{vec_grade_lengths()} returns a failing grade and informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{length}: \code{object} and \code{expected} are one-dimensional vectors of
different lengths
}
}

\examples{
.result <- 1:10
.solution <- 1:5
vec_check_length()
vec_grade_length()
}
