% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{vec_check_vector}
\alias{vec_check_vector}
\alias{vec_grade_vector}
\title{Check that the rows and columns of two tables are the same}
\usage{
vec_check_vector(
  object = .result,
  expected = .solution,
  check_class = TRUE,
  check_length = TRUE,
  check_levels = TRUE,
  check_values = TRUE,
  check_names = TRUE,
  env = parent.frame()
)

vec_grade_vector(
  object = .result,
  expected = .solution,
  max_diffs = 3,
  check_class = TRUE,
  check_length = TRUE,
  check_values = TRUE,
  check_names = TRUE,
  env = parent.frame(),
  ...
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{vec_check_vector()} and \code{vec_grade_vector()} were renamed to \code{\link[=vec_check]{vec_check()}}
and \code{\link[=vec_grade]{vec_grade()}}.
}
\keyword{internal}
