% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{check_names}
\alias{check_names}
\title{Check that the names of two object are the same}
\usage{
check_names(object, expected, max_print = 3)
}
\arguments{
\item{object}{An object to be compared to \code{expected}.}

\item{expected}{An object with the expected names.}

\item{max_print}{The maximum number of missing and/or unexpected names to include
in an informative failure message. Defaults to 3.}
}
\value{
Returns \code{object} invisibly.
}
\description{
Checks if \code{object} and \code{expected} have the same \link[=names]{names}.
If the names differ, returns a failure state and an informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}.
}
\section{Problem}{

\enumerate{
\item \code{names}: The table has names that are not expected,
or is missing names that are expected.
}
}

