% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_table.R
\name{check_table}
\alias{check_table}
\title{Check that the rows and columns of two tables are the same}
\usage{
check_table(
  object = .result,
  expected = .solution,
  max_print = 3,
  check_nrow = TRUE,
  check_names = TRUE,
  check_ncol = !check_names,
  check_columns = TRUE,
  check_class = check_columns,
  check_values = check_columns
)
}
\arguments{
\item{object}{A data frame to be compared to \code{expected}.}

\item{expected}{A data frame containing the expected result.}

\item{max_print}{The maximum number of mismatched values to display in an
informative failure message.
Passed to \code{\link[=check_names]{check_names()}} to determine the number of mismatched column
names to display and the \code{n_values} argument of \code{\link[=check_column]{check_column()}} to
determine the number of mismatched column values to display.
Defaults to 3.}

\item{check_nrow}{A \link{logical} indicating whether to check that \code{object} and
\code{expected} have the same number of rows with \code{\link[=nrow]{nrow()}}.}

\item{check_names}{A \link{logical} indicating whether to check that \code{object} and
\code{expected} have the same column names with \code{\link[=check_names]{check_names()}}.}

\item{check_ncol}{A \link{logical} indicating whether to check that \code{object} and
\code{expected} have the same number of columns with \code{\link[=ncol]{ncol()}}.
By default, \code{check_ncol} is \code{\link{FALSE}} if \code{check_names} is \code{\link{TRUE}}, as
running both is redundant.}

\item{check_columns}{A \link{logical} indicating whether to check that all columns
have the same contents with \code{\link[=check_column]{check_column()}}.}

\item{check_class}{A logical indicating whether to check that \code{name} has the
same class in \code{object} and \code{expected}.}

\item{check_values}{A logical indicating whether to check that \code{name} has the
same values in \code{object} and \code{expected}.}
}
\value{
Returns \code{object} invisibly.
}
\description{
Checks if \code{object} and \code{expected} have the same number of rows, the same
column names, and the same column contents.
If the data frames differ, returns a failure state and an informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}.
}
