% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vector.R
\name{check_vector}
\alias{check_vector}
\title{Checks that two vectors are the same}
\usage{
check_vector(
  object = .result,
  expected = .solution,
  max_diffs = 3,
  check_class = TRUE,
  check_length = TRUE,
  check_values = TRUE,
  unit = "result",
  prefix = "vector_"
)
}
\arguments{
\item{object}{A vector to be compared to \code{expected}.}

\item{expected}{A vector containing the expected result.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of mismatched values to
print. Defaults to 3.}

\item{check_class}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same classes.}

\item{check_length}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same length.}

\item{check_values}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} contain the same values.}

\item{unit}{\verb{[character(1)]}\cr The label used to describe the vector in
feedback messages. Defaults to \code{"result"}.}

\item{problem_prefix}{\verb{[character(1)]}\cr The prefix appended to the
\code{problem} label in \code{\link[gradethis:graded]{gradethis::fail()}} objects. Defaults to \code{"vector_"}.}
}
\value{
If there are any issues, generate a \code{\link[gradethis:graded]{gradethis::fail()}} message.
Otherwise, invisibly return \code{\link{NULL}}.
}
\description{
Check if two vectors have the same class, length, and values.
If the vectors differ, returns a failure state and an informative
message with \code{\link[gradethis:graded]{gradethis::fail()}}.
}
\section{Problems}{

\enumerate{
\item \code{vector_class}: \code{object} doesn't have the same classes as \code{expected}
\item \code{vector_length}: \code{object} doesn't have the same length as \code{expected}
\item \code{vector_values}: \code{object} doesn't contain the same values as \code{expected}
}
}

