% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_check_class-friendly_class.R
\name{friendly_class}
\alias{friendly_class}
\alias{friendly_class,ANY-method}
\alias{friendly_class,character-method}
\alias{friendly_class,numeric-method}
\alias{friendly_class,integer-method}
\alias{friendly_class,logical-method}
\alias{friendly_class,complex-method}
\alias{friendly_class,raw-method}
\alias{friendly_class,factor-method}
\alias{friendly_class,Date-method}
\alias{friendly_class,POSIXt-method}
\alias{friendly_class,Period-method}
\alias{friendly_class,data.frame-method}
\alias{friendly_class,tbl_df-method}
\alias{friendly_class,grouped_df-method}
\alias{friendly_class,rowwise_df-method}
\alias{friendly_class,list-method}
\alias{friendly_class,matrix-method}
\alias{friendly_class,array-method}
\title{Generate a human-readable description of an object's class}
\usage{
friendly_class(object)

\S4method{friendly_class}{ANY}(object)

\S4method{friendly_class}{character}(object)

\S4method{friendly_class}{numeric}(object)

\S4method{friendly_class}{integer}(object)

\S4method{friendly_class}{logical}(object)

\S4method{friendly_class}{complex}(object)

\S4method{friendly_class}{raw}(object)

\S4method{friendly_class}{factor}(object)

\S4method{friendly_class}{Date}(object)

\S4method{friendly_class}{POSIXt}(object)

\S4method{friendly_class}{Period}(object)

\S4method{friendly_class}{data.frame}(object)

\S4method{friendly_class}{tbl_df}(object)

\S4method{friendly_class}{grouped_df}(object)

\S4method{friendly_class}{rowwise_df}(object)

\S4method{friendly_class}{list}(object)

\S4method{friendly_class}{matrix}(object)

\S4method{friendly_class}{array}(object)
}
\arguments{
\item{object}{An object whose class will be described}
}
\value{
A \link{character} string of length 1,
based on the \link{class} and \link{length} of \code{object}.
}
\description{
Generate a human-readable description of an object's class
}
