% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_this.R
\name{grade_this_table}
\alias{grade_this_table}
\title{Grade this table}
\usage{
grade_this_table(
  correct = NULL,
  pre_check = NULL,
  post_check = NULL,
  pass_if_equal = TRUE,
  ...,
  max_diffs = 3,
  cols = NULL,
  check_class = TRUE,
  ignore_class = NULL,
  check_names = TRUE,
  check_column_order = FALSE,
  check_dimensions = TRUE,
  check_groups = TRUE,
  check_columns = TRUE,
  check_column_class = check_columns,
  check_column_values = check_columns,
  tolerance = sqrt(.Machine$double.eps),
  hint = getOption("gradethis.fail.hint", FALSE),
  encourage = getOption("gradethis.fail.encourage", FALSE),
  pass.praise = NULL
)
}
\arguments{
\item{correct}{\verb{[character(1)]}\cr The message shown to the student when
their \code{.result} matches the exercise \code{.solution}, if \code{pass_if_equal} is
\code{TRUE}.}

\item{pre_check, post_check}{\verb{[expression]}\cr Code to run before or after the
table or vector grading is performed. The pre check runs before calling
\code{\link[gradethis:pass_if_equal]{gradethis::pass_if_equal()}} so that you can modify or adjust the student's
\code{.result} or the \code{.solution} if there are parts of either that need to be
ignored. These arguments can also be used in conjunction with the
\code{pass_if_equal} option when the grading requirements are more involved.}

\item{pass_if_equal}{\verb{[logical(1)]}\cr When \code{TRUE} (default), the \code{.result}
is compared to the \code{.solution} with \code{\link[gradethis:pass_if_equal]{gradethis::pass_if_equal()}} after the
\emph{pre check} and before calling the \pkg{tblcheck} grading function.}

\item{...}{Additional arguments passed to \code{graded()} or additional data to be
included in the feedback object.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of mismatched values to
display in an informative failure message.
Passed to \code{\link[=tbl_check_names]{tbl_check_names()}} to determine the number of mismatched column
names to display and the \code{n_values} argument of \code{\link[=tbl_check_column]{tbl_check_column()}} to
determine the number of mismatched column values to display.
Defaults to 3.}

\item{cols}{[\code{\link[tidyselect:language]{tidy-select}}]\cr A selection of columns
to compare between \code{object} and \code{expected}. Differences in other columns
will be ignored. If \code{\link{NULL}}, the default, all columns will be checked.}

\item{check_class}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same classes with \code{\link[=tbl_check_class]{tbl_check_class()}}.}

\item{ignore_class}{\verb{[character()]}\cr A vector of classes to ignore when
finding differences between \code{object} and \code{expected}.

If an element is named, differences will only be ignored between the pair
of the element and its name.
For example, \code{ignore_class = c("integer" = "numeric")} will ignore class
differences only if \code{object} has class \link{integer} and \code{expected} has class
\link{numeric}, or vice versa.

If all the classes of \code{expected} are included in \code{ignore_class},
a \code{class} problem will never be returned.}

\item{check_names}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same column names with \code{\link[=tbl_check_names]{tbl_check_names()}}.}

\item{check_column_order}{\verb{[logical(1)]}\cr Whether to check that the columns
of  \code{object} are in the same order as \code{expected} with \code{\link[=tbl_check_names]{tbl_check_names()}}.
Defaults to \code{FALSE}.}

\item{check_dimensions}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same number of rows and columns
with \code{\link[=tbl_check_dimensions]{tbl_check_dimensions()}}.}

\item{check_groups}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same \link[dplyr:group_by]{groups}
with \code{\link[dplyr:group_data]{dplyr::group_vars()}}.}

\item{check_columns}{\verb{[logical(1)]}\cr Whether to check that all columns
have the same contents with \code{\link[=tbl_check_column]{tbl_check_column()}}.}

\item{check_column_class}{\verb{[logical(1)]}\cr Whether to check that each
columns has the same class in \code{object} and \code{expected}.}

\item{check_column_values}{\verb{[logical(1)]}\cr Whether to check that each
column has the same values in \code{object} and \code{expected}.}

\item{tolerance}{\verb{[numeric(1) ≥ 0]}\cr \code{values} differences smaller than
\code{tolerance} are ignored. The default value is close to \code{1.5e-8}.}

\item{hint}{Include a code feedback hint with the failing message? This
argument only applies to \code{fail()} and \code{fail_if_equal()} and the message is
added using the default options of \code{\link[gradethis:give_code_feedback]{give_code_feedback()}} and
\code{\link[gradethis:maybe_code_feedback]{maybe_code_feedback()}}. The default value of \code{hint} can be set using
\code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.hint} option.}

\item{encourage}{Include a random encouraging phrase with
\code{\link[gradethis:random_encouragement]{random_encouragement()}}? The default value of \code{encourage} can be set
using \code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.encourage} option.}

\item{pass.praise}{Logical \code{TRUE} or \code{FALSE} to determine whether a praising
phrase should be automatically prepended to any \code{\link[gradethis:pass]{pass()}} or
\code{\link[gradethis:pass_if_equal]{pass_if_equal()}} messages. Sets the \code{gradethis.pass.praise} option.}
}
\value{
The returned feedback is equivalent to \pkg{gradethis} grading code
using \code{\link[gradethis:grade_this]{grade_this()}} with the following
components:
\enumerate{
\item First the \code{pre_check} code, if any, is evaluated. If this code calls
\code{\link[gradethis:graded]{pass()}}, \code{\link[gradethis:graded]{fail()}}, or their
equivalents, that feedback is provided immediately.
\item If \code{pass_if_equal} is \code{TRUE}, then
\code{\link[gradethis:pass_if_equal]{pass_if_equal()}} is called to compare the
\code{\link[gradethis:grade_this-objects]{.result}} to the
\code{\link[gradethis:grade_this-objects]{.solution}}. The message in \code{correct}
is used for the feedback.
\item The appropriate \pkg{tblcheck} grading function is called, returning
any feedback:
\enumerate{
\item \code{grade_this_table()} returns the results from \code{\link[=tbl_grade]{tbl_grade()}}
\item \code{grade_this_vector()} returns the results from \code{\link[=vec_grade]{vec_grade()}}
}
\item The \code{post_check} code, if any, is evaluated and any feedback from a call
to \code{\link[gradethis:graded]{pass()}}, \code{\link[gradethis:graded]{fail()}}, or
their equivalents is returned.
\item Finally, if no other feedback is returned, the feedback from
\code{\link[gradethis:graded]{gradethis::fail()}} is provided to the student, using the options
\code{fail.message}, \code{fail.hint} and \code{fail.encourage}.
}
}
\description{
Automatically grade a table resulting from student code using
\code{\link[gradethis:grade_this]{gradethis::grade_this()}} and \code{\link[=tbl_grade]{tbl_grade()}} to compare the
student's result with the author's solution.
}
\examples{
ex <- gradethis::mock_this_exercise(
  .solution_code = tibble::tibble(x = 1:3, y = letters[x]),
  .user_code = tibble::tibble(x = 1:3, y = c("A", "b", "c"))
)

## Grading Tables ----
grade_this_table()(ex)

# Roughly equivalent to...
gradethis::grade_this({
  gradethis::pass_if_equal()
  tbl_grade()
  gradethis::fail()
})(ex)
}
\seealso{
\code{\link[=tbl_grade]{tbl_grade()}}

Other graders: 
\code{\link{grade_this_vector}()}
}
\concept{graders}
