% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{max_setdiff}
\alias{max_setdiff}
\title{Find up to a certain number of differences between two vectors}
\usage{
max_setdiff(x, y, max = Inf)
}
\arguments{
\item{x, y}{Vectors. Elements that appear in \code{x} but not \code{y} will be listed.}

\item{max}{The maximum number of elements to list.
Defaults to \link{Inf}, which includes all elements.}
}
\value{
A vector with no more than \code{max} elements.
}
\description{
Find up to a certain number of differences between two vectors
}
\keyword{internal}
