% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_message.R
\name{problem_message}
\alias{problem_message}
\title{Create a message from a problem object}
\usage{
problem_message(problem, ...)
}
\arguments{
\item{problem}{An object with base class \code{gradethis_problem}. Problems
identified by \pkg{tblcheck} also include \code{tblcheck_problem}, plus
additional classes that more specifically identify the problem type.}

\item{...}{Additional arguments passed to the underlying methods.}
}
\value{
A length-1 character string with a message describing the problem.
}
\description{
\code{problem_message()} is an S3 generic that powers the conversion of problems
detected by \code{\link[=tbl_check]{tbl_check()}} and \code{\link[=vec_check]{vec_check()}} (and their related helper
functions) into a human-readable message.
}
\examples{
problem <- problem(
  type = "class",
  expected = "character",
  actual = "numeric",
  expected_length = 1,
  actual_length = 2
)

problem_message(problem)

}
\seealso{
Other Problem functions: 
\code{\link{problem_grade}()},
\code{\link{problem_type}()},
\code{\link{problem}()}
}
\concept{Problem functions}
