% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{problem_type}
\alias{problem_type}
\alias{is_problem}
\alias{is_tblcheck_problem}
\alias{as_problem}
\title{Problem helper functions}
\usage{
problem_type(x)

is_problem(x, type = NULL)

is_tblcheck_problem(x, type = NULL)

as_problem(x)
}
\arguments{
\item{x}{An object}

\item{type}{\verb{[character(1)]}\cr A \code{problem} type}
}
\value{
\code{is_problem()} and \code{is_tblcheck_problem()} return a \link{logical}
of length 1.
\code{problem_type()} returns a \link{character} of length 1.
\code{as_problem()} returns a \code{tblcheck_problem}.
}
\description{
\itemize{
\item \code{problem_type()} returns a problem's type, or \code{\link{NULL}} if the input is
not a problem.
\item \code{is_problem()} tests whether an object is a \code{gradethis} problem.
\item \code{is_tblcheck_problem()} tests whether an object is a problem created
by \code{tblcheck}.
\item \code{as_problem()} converts a list to a \code{tblcheck_problem}.
}

If \code{type} is specified, \code{is_problem()} and \code{is_tblcheck_problem()} test
whether an object is a problem of the specified type.
}
\examples{
problem_type(tbl_check_vector(1, "1"))
is_problem(tbl_check_vector(1, "1"), "vector_class")
is_tblcheck_problem(tbl_check_vector(1, "1"), "class")
}
