% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_class.R
\name{tbl_check_class}
\alias{tbl_check_class}
\alias{tbl_grade_class}
\title{Checks that two objects have the same classes}
\usage{
tbl_check_class(
  object = .result,
  expected = .solution,
  object_label = NULL,
  problem_prefix = "",
  envir = parent.frame()
)

tbl_grade_class(
  object = .result,
  expected = .solution,
  object_label = "result",
  problem_prefix = "",
  envir = parent.frame()
)
}
\arguments{
\item{object}{An object to be compared to \code{expected}.}

\item{expected}{An object containing the expected result.}

\item{object_label}{\verb{[character(1)]}\cr The label used to describe the object
in feedback messages. Defaults to \code{"result"}.}

\item{problem_prefix}{\verb{[character(1)]}\cr The prefix appended to the
\code{problem} label in \code{\link[gradethis:graded]{gradethis::fail()}} objects.
Defaults to \code{""}, which appends no prefix.}

\item{envir}{The environment in which to find \code{.result} and \code{.solution}.}
}
\value{
If there are any issues, a \link{list} from \code{tbl_check_class()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{tbl_grade_class()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks if \code{object} and \code{expected} have the same \link[=class]{class}.
If the classes differ
\itemize{
\item \code{tbl_check_class()} returns a list describing the problem
\item \code{tbl_grade_class()} returns a failing grade and informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{class}: The object does not have the expected classes
}
}

