% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_column.R
\name{tbl_check_column}
\alias{tbl_check_column}
\alias{tbl_grade_column}
\title{Checks that a column is identical across two tables}
\usage{
tbl_check_column(
  name,
  object = .result,
  expected = .solution,
  max_diffs = 3,
  check_class = TRUE,
  check_length = TRUE,
  check_values = TRUE,
  envir = parent.frame()
)

tbl_grade_column(
  name,
  object = .result,
  expected = .solution,
  max_diffs = 3,
  check_class = TRUE,
  check_length = TRUE,
  check_values = TRUE,
  envir = parent.frame()
)
}
\arguments{
\item{name}{\verb{[character(1)]}\cr The name of the column to check.}

\item{object}{A data frame to be compared to \code{expected}.}

\item{expected}{A data frame containing the expected result.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of mismatched values to
print. Defaults to 3.}

\item{check_class}{\verb{[logical(1)]}\cr Whether to check that \code{name} has the
same class in \code{object} and \code{expected}.}

\item{check_length}{\verb{[logical(1)]}\cr Whether to check that \code{name} has the
same length in \code{object} and \code{expected}.}

\item{check_values}{\verb{[logical(1)]}\cr Whether to check that \code{name} has the
same values in \code{object} and \code{expected}.}

\item{envir}{The environment in which to find \code{.result} and \code{.solution}.}
}
\value{
If there are any issues, a \link{list} from \code{tbl_check_column()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{tbl_grade_column()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks for differences between the \code{name} column in \code{object} and in
\code{expected} in the following order:
\enumerate{
\item Check that the \code{name} column exists in \code{object}
\item Check class with \code{\link[=tbl_check_class]{tbl_check_class()}}
\item Check length with \code{\link[=tbl_check_dimensions]{tbl_check_dimensions()}}
\item Check column values with \code{\link[=tbl_check_vector]{tbl_check_vector()}}
}

If the columns differ
\itemize{
\item \code{tbl_check_column()} returns a list describing the problem
\item \code{tbl_grade_column()} returns a failing grade and informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{column_class}: Any mismatch in the classes of the \code{name} column
\item \code{column_dimensions}: The \code{name} column doesn't have the expected length
\item \code{column_values}: The \code{name} column doesn't have the expected values
\item \code{column}: The \code{name} column doesn't appear in the \code{object}
}
}

\examples{
.result <- tibble::tibble(a = 1:10, b = 11:20)
.solution <- tibble::tibble(a = letters[1:10], b = letters[11:20])
tbl_check_column("a")
tbl_grade_column("a")

.result <- tibble::tibble(a = 1:10, b = 11:20)
.solution <- tibble::tibble(a = 1:11, b = 12:22)
tbl_check_column("a")
tbl_grade_column("a")

.result <- tibble::tibble(a = 1:10, b = 11:20)
.solution <- tibble::tibble(a = 11:20, b = 1:10)
tbl_check_column("a")
tbl_grade_column("a")
tbl_grade_column("a", max_diffs = 5)
tbl_grade_column("a", max_diffs = Inf)
}
