% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_column.R
\name{tbl_check_column}
\alias{tbl_check_column}
\alias{tbl_grade_column}
\title{Checks that a column is identical across two tables}
\usage{
tbl_check_column(
  column,
  object = .result,
  expected = .solution,
  check_class = TRUE,
  check_length = TRUE,
  check_values = TRUE,
  check_names = FALSE,
  env = parent.frame()
)

tbl_grade_column(
  column,
  object = .result,
  expected = .solution,
  max_diffs = 3,
  check_class = TRUE,
  check_length = TRUE,
  check_values = TRUE,
  check_names = FALSE,
  env = parent.frame(),
  ...
)
}
\arguments{
\item{column}{\verb{[character(1)]}\cr The name of the column to check.}

\item{object}{A data frame to be compared to \code{expected}.}

\item{expected}{A data frame containing the expected result.}

\item{check_class}{\verb{[logical(1)]}\cr Whether to check that \code{column} has the
same class in \code{object} and \code{expected}.}

\item{check_length}{\verb{[logical(1)]}\cr Whether to check that \code{column} has the
same length in \code{object} and \code{expected}.}

\item{check_values}{\verb{[logical(1)]}\cr Whether to check that \code{column} has the
same values in \code{object} and \code{expected}.}

\item{check_names}{\verb{[logical(1)]}\cr Whether to check that \code{column} has the
same \link[=names]{names} in \code{object} and \code{expected}.
Defaults to \code{FALSE}.}

\item{env}{The environment in which to find \code{.result} and \code{.solution}.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of mismatched values to
print. Defaults to 3.}

\item{...}{Additional arguments passed to \code{\link[gradethis:graded]{gradethis::fail()}}.}
}
\value{
If there are any issues, a \link{list} from \code{tbl_check_column()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{tbl_grade_column()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks if \code{column} has the same class and values in \code{object} and \code{expected}.
If the columns differ
\itemize{
\item \code{tbl_check_column()} returns a list describing the problem
\item \code{tbl_grade_column()} returns a failing grade and informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{names} (\code{table_problem}): \code{object} doesn't contain a column named column.
\item \code{class}: Any mismatch in the classes of the \code{column}.
\item \code{length}: The \code{column} doesn't have the expected length.
\item \code{levels_n}, \code{levels}, \code{levels_reversed}, \code{levels_order}:
See \code{\link[=vec_check_levels]{vec_check_levels()}}.
\item \code{values}: The \code{column} doesn't have the expected values.
\item \code{names} (\code{column_problem}): The \code{column} has different \link[=names]{names}
than expected.
\item \code{names_order}: The \code{column} has the same \link[=names]{names} as expected,
but in a different order.
}
}

\examples{
.result <- tibble::tibble(a = 1:10, b = 11:20)
.solution <- tibble::tibble(a = letters[1:10], b = letters[11:20])
tbl_check_column("a")
tbl_grade_column("a")

.result <- tibble::tibble(a = 1:10, b = 11:20)
.solution <- tibble::tibble(a = 1:11, b = 12:22)
tbl_check_column("a")
tbl_grade_column("a")

.result <- tibble::tibble(a = 1:10, b = 11:20)
.solution <- tibble::tibble(a = 11:20, b = 1:10)
tbl_check_column("a")
tbl_grade_column("a")
tbl_grade_column("a", max_diffs = 5)
tbl_grade_column("a", max_diffs = Inf)
}
