% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dimensions.R
\name{tbl_check_dimensions}
\alias{tbl_check_dimensions}
\alias{tbl_grade_dimensions}
\title{Check that the length of two object are the same}
\usage{
tbl_check_dimensions(
  object = .result,
  expected = .solution,
  envir = parent.frame()
)

tbl_grade_dimensions(
  object = .result,
  expected = .solution,
  envir = parent.frame()
)
}
\arguments{
\item{object}{An object to be compared to \code{expected}.}

\item{expected}{An object containing the expected result.}

\item{envir}{The environment in which to find \code{.result} and \code{.solution}.}
}
\value{
If there are any issues, a \link{list} from \code{tbl_check_dimensions()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{tbl_grade_dimensions()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks if \code{object} and \code{expected} have the same \link[=dim]{dimenisons}.
If the names differ
\itemize{
\item \code{tbl_check_dimensions()} returns a list describing the problem
\item \code{tbl_grade_dimensions()} returns a failing grade and informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{dimensions}: \code{object} and \code{expected} have different dimensions
}
}

