% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dimensions.R
\name{tbl_check_dimensions}
\alias{tbl_check_dimensions}
\alias{tbl_grade_dimensions}
\title{Check that the dimensions of two object are the same}
\usage{
tbl_check_dimensions(
  object = .result,
  expected = .solution,
  env = parent.frame()
)

tbl_grade_dimensions(
  object = .result,
  expected = .solution,
  env = parent.frame()
)
}
\arguments{
\item{object}{An object to be compared to \code{expected}.}

\item{expected}{An object containing the expected result.}

\item{env}{The environment in which to find \code{.result} and \code{.solution}.}
}
\value{
If there are any issues, a \link{list} from \code{tbl_check_dimensions()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{tbl_grade_dimensions()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks if \code{object} and \code{expected} have the same \link[=dim]{dimenisons}.
If the dimensions differ
\itemize{
\item \code{tbl_check_dimensions()} returns a list describing the problem
\item \code{tbl_grade_dimensions()} returns a failing grade and informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{n_dimensions}: \code{object} and \code{expected} have a different number
of dimensions
\item \code{length}: \code{object} and \code{expected} are one-dimensional vectors of
different lengths
\item \code{ncol}: \code{object} and \code{expected} are two-dimensional objects with a
different number of columns
\item \code{nrow}: \code{object} and \code{expected} are two-dimensional objects with a
different number of rows
\item \code{dimensions}: \code{object} and \code{expected} are multi-dimensional arrays with
different dimensions
}
}

\examples{
.result <- 1:10
.solution <- 1:5
tbl_check_dimensions()
tbl_grade_dimensions()

.result <- tibble::tibble(a = 1:10, b = 1:10, c = 1:10)
.solution <- tibble::tibble(a = 1:10, b = 1:10)
tbl_check_dimensions()
tbl_grade_dimensions()

.result <- tibble::tibble(a = 1:10, b = 1:10)
.solution <- tibble::tibble(a = 1:5, b = 1:5)
tbl_check_dimensions()
tbl_grade_dimensions()

.result <- 1:12
.solution <- matrix(1:12, 3)
tbl_check_dimensions()
tbl_grade_dimensions()
}
