% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_groups.R
\name{tbl_check_groups}
\alias{tbl_check_groups}
\alias{tbl_grade_groups}
\title{Check that the groups of two object are the same}
\usage{
tbl_check_groups(object = .result, expected = .solution, env = parent.frame())

tbl_grade_groups(
  object = .result,
  expected = .solution,
  max_diffs = 3,
  env = parent.frame()
)
}
\arguments{
\item{object}{An object to be compared to \code{expected}.}

\item{expected}{An object containing the expected result.}

\item{env}{The environment in which to find \code{.result} and \code{.solution}.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of missing and/or
unexpected names to include in an informative failure message.
Defaults to 3.}
}
\value{
If there are any issues, a \link{list} from \code{tbl_check_groups()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{tbl_grade_groups()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks if \code{object} and \code{expected} have the same \link[dplyr:group_by]{groups}.
If the groups differ
\itemize{
\item \code{tbl_check_groups()} returns a list describing the problem
\item \code{tbl_grade_groups()} returns a failing grade and informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{groups}: The object has groups that are not expected,
or is missing groups that are expected.
}
}

\examples{
.result <- dplyr::group_by(tibble::tibble(a = 1:10, b = 11:20), a)
.solution <- dplyr::group_by(tibble::tibble(a = 1:10, b = 11:20), b)
tbl_check_groups()
tbl_grade_groups()
}
