% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_is_table.R
\name{tbl_check_is_table}
\alias{tbl_check_is_table}
\alias{tbl_grade_is_table}
\title{Checks that an object is a table}
\usage{
tbl_check_is_table(object = .result, env = parent.frame())

tbl_grade_is_table(object = .result, env = parent.frame(), ...)
}
\arguments{
\item{object}{An object to be compared to \code{expected}.}

\item{env}{The environment in which to find \code{.result} and \code{.solution}.}

\item{...}{
  Arguments passed on to \code{\link[gradethis:graded]{gradethis::fail}}
  \describe{
    \item{\code{hint}}{Include a code feedback hint with the failing message? This
argument only applies to \code{fail()} and \code{fail_if_equal()} and the message is
added using the default options of \code{\link[gradethis:code_feedback]{give_code_feedback()}} and
\code{\link[gradethis:code_feedback]{maybe_code_feedback()}}. The default value of \code{hint} can be set using
\code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.hint} option.}
    \item{\code{encourage}}{Incude a random encouraging phrase with
\code{\link[gradethis:praise]{random_encouragement()}}? The default value of \code{encourage} can be set
using \code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.encourage} option.}
  }}
}
\value{
If there are any issues, a \link{list} from \code{tbl_check_is_table()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{tbl_grade_is_table()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks if \code{object} inherits the \link{data.frame} class.
If the not
\itemize{
\item \code{tbl_check_is_table()} returns a list describing the problem
\item \code{tbl_grade_is_table()} returns a failing grade and
informative message with \code{\link[gradethis:graded]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{not_table}: The object is not a table
}
}

\examples{
.result <- data.frame(a = 1:10)
tbl_check_is_table()
tbl_grade_is_table()

.result <- tibble::tibble(a = 1:10)
tbl_check_is_table()
tbl_grade_is_table()

.result <- list(a = 1:10)
tbl_check_is_table()
tbl_grade_is_table()

}
