% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{tbl_check_names}
\alias{tbl_check_names}
\alias{tbl_grade_names}
\title{Check that the names of two object are the same}
\usage{
tbl_check_names(
  object = .result,
  expected = .solution,
  object_label = NULL,
  problem_prefix = "",
  envir = parent.frame()
)

tbl_grade_names(
  object = .result,
  expected = .solution,
  max_diffs = 3,
  envir = parent.frame()
)
}
\arguments{
\item{object}{An object to be compared to \code{expected}.}

\item{expected}{An object containing the expected result.}

\item{object_label}{\verb{[character(1)]}\cr The label used to describe the object
in feedback messages. Defaults to \code{"result"}.}

\item{problem_prefix}{\verb{[character(1)]}\cr The prefix appended to the
\code{problem} label in \code{\link[gradethis:graded]{gradethis::fail()}} objects.
Defaults to \code{""}, which appends no prefix.}

\item{envir}{The environment in which to find \code{.result} and \code{.solution}.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of missing and/or
unexpected names to include in an informative failure message.
Defaults to 3.}
}
\value{
If there are any issues, a \link{list} from \code{tbl_check_names()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{tbl_grade_names()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks if \code{object} and \code{expected} have the same \link[=names]{names}.
If the names differ
\itemize{
\item \code{tbl_check_names()} returns a list describing the problem
\item \code{tbl_grade_names()} returns a failing grade and informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{names}: The object has names that are not expected,
or is missing names that are expected.
}
}

