% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nchar.R
\name{tbl_check_nchar}
\alias{tbl_check_nchar}
\alias{tbl_grade_nchar}
\title{Check that the two vectors have all the same number of characters}
\usage{
tbl_check_nchar(object = .result, expected = .solution, envir = parent.frame())

tbl_grade_nchar(object = .result, expected = .solution, envir = parent.frame())
}
\arguments{
\item{object}{An object to be compared to \code{expected}.}

\item{expected}{An object containing the expected result.}

\item{envir}{The environment in which to find \code{.result} and \code{.solution}.}
}
\value{
If there are any issues, a \link{list} from \code{tbl_check_nchar()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{tbl_grade_nchar()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks if every element in \code{object} and \code{expected} have the same
\link[=nchar]{number of characters}.
If the number of characters differ
\itemize{
\item \code{tbl_check_nchar()} returns a list describing the problem
\item \code{tbl_grade_nchar()} returns a failing grade and informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}
}

Every element of \code{expected} must have the same number of characters.
If \code{expected} does not have a consistent number of characters, no grade will
be produced.

If \code{object} and \code{expected} are not both \link{character} vectors, they will be
passed to \code{\link[=tbl_check_class]{tbl_check_class()}} for grading.
}
\section{Problems}{

\enumerate{
\item \code{inconsistent_nchar}: every element of \code{object} does not have the same
number of characters
\item \code{nchar}: every element of \code{object} does not have the same number of
characters as \code{expected}
}

Additional problems may be produced by \code{\link[=tbl_check_class]{tbl_check_class()}}
}

\examples{
.result <- letters
.solution <- strrep(letters, 2)
tbl_check_nchar()
tbl_grade_nchar()

.result <- strrep(letters, 1:26)
.solution <- strrep(letters, 2)
tbl_check_nchar()
tbl_grade_nchar()

.result <- 1:26
.solution <- strrep(letters, 2)
tbl_check_nchar()
tbl_grade_nchar()
}
