% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_table.R
\name{tbl_check_table}
\alias{tbl_check_table}
\alias{tbl_grade_table}
\title{Check that the rows and columns of two tables are the same}
\usage{
tbl_check_table(
  object = .result,
  expected = .solution,
  check_class = TRUE,
  check_names = TRUE,
  check_column_order = FALSE,
  check_dimensions = TRUE,
  check_groups = TRUE,
  check_columns = TRUE,
  check_column_class = check_columns,
  check_column_values = check_columns,
  env = parent.frame()
)

tbl_grade_table(
  object = .result,
  expected = .solution,
  max_diffs = 3,
  check_class = TRUE,
  check_names = TRUE,
  check_column_order = FALSE,
  check_dimensions = TRUE,
  check_groups = TRUE,
  check_columns = TRUE,
  check_column_class = check_columns,
  check_column_values = check_columns,
  env = parent.frame(),
  ...
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{tbl_check_table()} and \code{tbl_grade_table()} were renamed to \code{\link[=tbl_check]{tbl_check()}}
and \code{\link[=tbl_grade]{tbl_grade()}}.
}
\keyword{internal}
