% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vector.R
\name{tbl_check_vector}
\alias{tbl_check_vector}
\alias{tbl_grade_vector}
\title{Checks that two vectors are the same}
\usage{
tbl_check_vector(
  object = .result,
  expected = .solution,
  max_diffs = 3,
  check_class = TRUE,
  check_length = TRUE,
  check_values = TRUE,
  check_names = TRUE,
  envir = parent.frame()
)

tbl_grade_vector(
  object = .result,
  expected = .solution,
  max_diffs = 3,
  check_class = TRUE,
  check_length = TRUE,
  check_values = TRUE,
  check_names = TRUE,
  envir = parent.frame()
)
}
\arguments{
\item{object}{A vector to be compared to \code{expected}.}

\item{expected}{A vector containing the expected result.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of mismatched values to
print. Defaults to 3.}

\item{check_class}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same classes.}

\item{check_length}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same length.}

\item{check_values}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} contain the same values.}

\item{check_names}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same names.}

\item{envir}{The environment in which to find \code{.result} and \code{.solution}.}
}
\value{
If there are any issues, a \link{list} from \code{tbl_check_vector()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{tbl_grade_vector()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Checks for differences between \code{object} and \code{expected} in the following order:
\enumerate{
\item Check class with \code{\link[=tbl_check_class]{tbl_check_class()}}
\item Check length with \code{\link[=tbl_check_dimensions]{tbl_check_dimensions()}}
\item Check vector values are the same
\item Check names with \code{\link[=tbl_check_names]{tbl_check_names()}}
}

If the vectors differ
\itemize{
\item \code{tbl_check_vector()} returns a list describing the problem
\item \code{tbl_grade_vector()} returns a failing grade and informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{vector_class}: \code{object} doesn't have the same classes as \code{expected}
\item \code{vector_dimensions}: \code{object} doesn't have the same length as \code{expected}
\item \code{vector_values}: \code{object} doesn't contain the same values as \code{expected}
\item \code{vector_names}: \code{object} has different \code{names} than \code{expected}
}
}

\examples{
.result <- 1:10
.solution <- letters[1:10]
tbl_check_vector()
tbl_grade_vector()

.result <- 1:10
.solution <- 1:11
tbl_check_vector()
tbl_grade_vector()

.result <- 1:10
.solution <- rlang::set_names(1:10, letters[1:10])
tbl_check_vector()
tbl_grade_vector()
tbl_grade_vector(max_diffs = 5)
tbl_grade_vector(max_diffs = Inf)

.result <- 1:10
.solution <- 11:20
tbl_check_vector()
tbl_grade_vector()
tbl_grade_vector(max_diffs = 5)
tbl_grade_vector(max_diffs = Inf)
}
