% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_equal.R
\name{tbl_equal}
\alias{tbl_equal}
\title{Check that the rows and columns of two tables are the same}
\usage{
tbl_equal(
  object = .result,
  expected = .solution,
  cols = NULL,
  check_class = TRUE,
  ignore_class = NULL,
  check_names = TRUE,
  check_column_order = FALSE,
  check_dimensions = TRUE,
  check_groups = TRUE,
  check_columns = TRUE,
  check_column_class = check_columns,
  check_column_levels = check_columns,
  check_column_values = check_columns,
  tolerance = sqrt(.Machine$double.eps),
  check_row_order = check_columns,
  env = parent.frame()
)
}
\arguments{
\item{object}{A data frame to be compared to \code{expected}.}

\item{expected}{A data frame containing the expected result.}

\item{cols}{[\code{\link[tidyselect:language]{tidy-select}}]\cr A selection of columns
to compare between \code{object} and \code{expected}. Differences in other columns
will be ignored. If \code{\link{NULL}}, the default, all columns will be checked.}

\item{check_class}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same classes with \code{\link[=tbl_check_class]{tbl_check_class()}}.}

\item{ignore_class}{\verb{[character()]}\cr A vector of classes to ignore when
finding differences between \code{object} and \code{expected}.

If an element is named, differences will only be ignored between the pair
of the element and its name.
For example, \code{ignore_class = c("integer" = "numeric")} will ignore class
differences only if \code{object} has class \link{integer} and \code{expected} has class
\link{numeric}, or vice versa.

If all the classes of \code{expected} are included in \code{ignore_class},
a \code{class} problem will never be returned.}

\item{check_names}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same column names with \code{\link[=tbl_check_names]{tbl_check_names()}}.}

\item{check_column_order}{\verb{[logical(1)]}\cr Whether to check that the columns
of  \code{object} are in the same order as \code{expected} with \code{\link[=tbl_check_names]{tbl_check_names()}}.
Defaults to \code{FALSE}.}

\item{check_dimensions}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same number of rows and columns
with \code{\link[=tbl_check_dimensions]{tbl_check_dimensions()}}.}

\item{check_groups}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same \link[dplyr:group_by]{groups}
with \code{\link[dplyr:group_data]{dplyr::group_vars()}}.}

\item{check_columns}{\verb{[logical(1)]}\cr Whether to check that all columns
have the same contents with \code{\link[=tbl_check_column]{tbl_check_column()}}.}

\item{check_column_class}{\verb{[logical(1)]}\cr Whether to check that each
column has the same class in \code{object} and \code{expected}.}

\item{check_column_levels}{\verb{[logical(1)]}\cr Whether to check that each
column has the same \link[=levels]{factor levels} in \code{object} and \code{expected}.}

\item{check_column_values}{\verb{[logical(1)]}\cr Whether to check that each
column has the same values in \code{object} and \code{expected}.}

\item{tolerance}{\verb{[numeric(1) ≥ 0]}\cr \code{values} differences smaller than
\code{tolerance} are ignored. The default value is close to \code{1.5e-8}.}

\item{check_row_order}{\verb{[logical(1)]}\cr Whether to check that the values in
each column are in the same order in \code{object} and \code{expected}.}

\item{env}{The environment in which to find \code{.result} and \code{.solution}.}
}
\value{
A \code{\link{TRUE}} or \code{\link{FALSE}} value.
}
\description{
Test if two tables are equivalent using the same process as \code{\link[=tbl_check]{tbl_check()}}.
Unlike \code{\link[=tbl_check]{tbl_check()}}, which returns either a \link{problem} object or \code{\link{NULL}},
\code{tbl_equal()} returns either \code{\link{TRUE}} or \code{\link{FALSE}}.
}
\examples{
tbl_equal(
	data.frame(a = 1:10, b = 11:20),
	data.frame(b = 11:20, a = 1:10)
)
}
