% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_check_values.R
\name{vec_check_values}
\alias{vec_check_values}
\alias{vec_grade_values}
\title{Checks that two vectors contain the same values}
\usage{
vec_check_values(
  object = .result,
  expected = .solution,
  tolerance = sqrt(.Machine$double.eps),
  env = parent.frame()
)

vec_grade_values(
  object = .result,
  expected = .solution,
  tolerance = sqrt(.Machine$double.eps),
  max_diffs = 3,
  env = parent.frame(),
  ...
)
}
\arguments{
\item{object}{A vector to be compared to \code{expected}.}

\item{expected}{A vector containing the expected result.}

\item{tolerance}{\verb{[numeric(1) ≥ 0]}\cr \code{values} differences smaller than
\code{tolerance} are ignored. The default value is close to \code{1.5e-8}.}

\item{env}{The environment in which to find \code{.result} and \code{.solution}.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of mismatched values to
print. Defaults to 3.}

\item{...}{
  Arguments passed on to \code{\link[gradethis:fail]{gradethis::fail}}
  \describe{
    \item{\code{correct}}{A logical value of whether or not the checked code is correct.}
    \item{\code{type,location}}{The \code{type} and \code{location} of the feedback object
provided to \pkg{learnr}. See
\url{https://rstudio.github.io/learnr/exercises.html#Custom_checking} for more
details.

\code{type} may be one of "auto", "success", "info", "warning", "error", or
"custom".

\code{location} may be one of "append", "prepend", or "replace".}
    \item{\code{praise}}{Include a random praising phrase with \code{\link[gradethis:random_praise]{random_praise()}}? The
default value of \code{praise} can be set using \code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the
\code{gradethis.pass.praise} option.}
    \item{\code{hint}}{Include a code feedback hint with the failing message? This
argument only applies to \code{fail()} and \code{fail_if_equal()} and the message is
added using the default options of \code{\link[gradethis:give_code_feedback]{give_code_feedback()}} and
\code{\link[gradethis:maybe_code_feedback]{maybe_code_feedback()}}. The default value of \code{hint} can be set using
\code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.hint} option.}
    \item{\code{encourage}}{Include a random encouraging phrase with
\code{\link[gradethis:random_encouragement]{random_encouragement()}}? The default value of \code{encourage} can be set
using \code{\link[gradethis:gradethis_setup]{gradethis_setup()}} or the \code{gradethis.fail.encourage} option.}
  }}
}
\value{
If there are any issues, a \link{list} from \code{vec_check_values()} or a
\code{\link[gradethis:fail]{gradethis::fail()}} message from \code{vec_grade_values()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Check if two vectors contain the same values.
If the values differ
\itemize{
\item \code{vec_check_values()} returns a list describing the problem
\item \code{vec_grade_values()} returns a failing grade and informative message
with \code{\link[gradethis:fail]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{values}: \code{object} doesn't contain the same values as \code{expected}
}
}

\examples{
.result <- 1:10
.solution <- letters[1:10]
vec_check_values()
vec_grade_values()

.result <- 1:10
.solution <- 1:11
vec_check_values()
vec_grade_values()

.result <- 1:10
.solution <- rlang::set_names(1:10, letters[1:10])
vec_check_values()
vec_grade_values()
vec_grade_values(max_diffs = 5)
vec_grade_values(max_diffs = Inf)

.result <- 1:10
.solution <- 11:20
vec_check_values()
vec_grade_values()
vec_grade_values(max_diffs = 5)
vec_grade_values(max_diffs = Inf)
}
