% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_values.R
\name{vec_check_values}
\alias{vec_check_values}
\alias{vec_grade_values}
\title{Checks that two vectors are contain the same values}
\usage{
vec_check_values(object = .result, expected = .solution, env = parent.frame())

vec_grade_values(
  object = .result,
  expected = .solution,
  max_diffs = 3,
  env = parent.frame()
)
}
\arguments{
\item{object}{A vector to be compared to \code{expected}.}

\item{expected}{A vector containing the expected result.}

\item{env}{The environment in which to find \code{.result} and \code{.solution}.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of mismatched values to
print. Defaults to 3.}
}
\value{
If there are any issues, a \link{list} from \code{vec_check_vector()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{vec_grade_vector()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Check if two vectors contain the same values.
If the values differ
\itemize{
\item \code{vec_check_values()} returns a list describing the problem
\item \code{vec_grade_values()} returns a failing grade and informative message
with \code{\link[gradethis:graded]{gradethis::fail()}}
}
}
\section{Problems}{

\enumerate{
\item \code{length}: \code{object} doesn't have the same length as \code{expected}.
\item \code{values}: \code{object} doesn't contain the same values as \code{expected}.
}
}

\examples{
.result <- 1:10
.solution <- letters[1:10]
vec_check_vector()
vec_grade_vector()

.result <- 1:10
.solution <- 1:11
vec_check_vector()
vec_grade_vector()

.result <- 1:10
.solution <- rlang::set_names(1:10, letters[1:10])
vec_check_vector()
vec_grade_vector()
vec_grade_vector(max_diffs = 5)
vec_grade_vector(max_diffs = Inf)

.result <- 1:10
.solution <- 11:20
vec_check_vector()
vec_grade_vector()
vec_grade_vector(max_diffs = 5)
vec_grade_vector(max_diffs = Inf)
}
