% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vector.R
\name{vec_check_vector}
\alias{vec_check_vector}
\alias{vec_grade_vector}
\title{Checks that two vectors are the same}
\usage{
vec_check_vector(
  object = .result,
  expected = .solution,
  max_diffs = 3,
  check_class = TRUE,
  check_length = TRUE,
  check_levels = TRUE,
  check_values = TRUE,
  check_names = TRUE,
  env = parent.frame()
)

vec_grade_vector(
  object = .result,
  expected = .solution,
  max_diffs = 3,
  check_class = TRUE,
  check_length = TRUE,
  check_values = TRUE,
  check_names = TRUE,
  env = parent.frame()
)
}
\arguments{
\item{object}{A vector to be compared to \code{expected}.}

\item{expected}{A vector containing the expected result.}

\item{max_diffs}{\verb{[numeric(1)]}\cr The maximum number of mismatched values to
print. Defaults to 3.}

\item{check_class}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same classes.}

\item{check_length}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same length.}

\item{check_levels}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same \link[=levels]{factor levels}.}

\item{check_values}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} contain the same values.}

\item{check_names}{\verb{[logical(1)]}\cr Whether to check that \code{object} and
\code{expected} have the same names.}

\item{env}{The environment in which to find \code{.result} and \code{.solution}.}
}
\value{
If there are any issues, a \link{list} from \code{vec_check_vector()} or a
\code{\link[gradethis:graded]{gradethis::fail()}} message from \code{vec_grade_vector()}.
Otherwise, invisibly returns \code{\link{NULL}}.
}
\description{
Check if two vectors have the same class, length, and values.
If the vectors differ, returns a failure state and an informative
message with \code{\link[gradethis:graded]{gradethis::fail()}}.
}
\section{Problems}{

\enumerate{
\item \code{vector_class}: \code{object} doesn't have the same classes as \code{expected}
\item \code{vector_length}: \code{object} doesn't have the same length as \code{expected}
\item \code{vector_n_levels}, \code{vector_levels}, \code{vector_level_order_diffs},
\code{vector_level_order}: See \code{\link[=vec_check_levels]{vec_check_levels()}}
\item \code{column_value_diffs}: The first \code{max_diffs} elements of \code{object} don't
contain the same values as \code{expected}
\item \code{vector_values}: \code{object} doesn't contain the same values as \code{expected}
\item \code{vector_names}: \code{object} has different \link[=names]{names} than \code{expected}
}
}

\examples{
.result <- 1:10
.solution <- letters[1:10]
vec_check_vector()
vec_grade_vector()

.result <- 1:10
.solution <- 1:11
vec_check_vector()
vec_grade_vector()

.result <- 1:10
.solution <- rlang::set_names(1:10, letters[1:10])
vec_check_vector()
vec_grade_vector()
vec_grade_vector(max_diffs = 5)
vec_grade_vector(max_diffs = Inf)

.result <- 1:10
.solution <- 11:20
vec_check_vector()
vec_grade_vector()
vec_grade_vector(max_diffs = 5)
vec_grade_vector(max_diffs = Inf)
}
